Move test_parse_load_args.py tests into test_config.py
